requirements.txt
fastapi
sqlalchemy
pymysql
passlib[bcrypt]
python-dotenv
pymongo  
httpx  
PyJWT  
uvicorn 
